+++ src/main/java/me/nazarxexe/survival/econ/EconomyAPI.java
@@ -2,7 +2,9 @@ package me.nazarxexe.survival.econ;
 
 import cn.nukkit.Player;
 import cn.nukkit.scheduler.AsyncTask;
+import cn.nukkit.utils.Hash;
 import cn.nukkit.utils.TextFormat;
+import lombok.Getter;
 import me.nazarxexe.survival.core.economy.EconomyManager;
 import me.nazarxexe.survival.core.economy.Pocket;
 import me.nazarxexe.survival.core.tools.TerminalComponent;
@@ -30,12 +32,13 @@ public class EconomyAPI {
         return cache;
     }
 
-
+    private @Getter HashMap<String , String> confighook;
 
     public EconomyAPI(@NotNull Economy plugin){
         this.plugin = plugin;
         this.manager = plugin.getEconomyManager();
         this.cache = new HashMap<>();
+        this.confighook = plugin.getConfighook();
     }
